refactor: rename to jsonutil
