+++ internal/handler/subcateg.go
@@ -6,7 +6,7 @@ import (
 	"github.com/OYE0303/expense-tracker-go/internal/domain"
 	"github.com/OYE0303/expense-tracker-go/pkg/ctxutil"
 	"github.com/OYE0303/expense-tracker-go/pkg/errutil"
-	"github.com/OYE0303/expense-tracker-go/pkg/jsutil"
+	"github.com/OYE0303/expense-tracker-go/pkg/jsonutil"
 	"github.com/OYE0303/expense-tracker-go/pkg/logger"
 	"github.com/OYE0303/expense-tracker-go/pkg/validator"
 )
@@ -26,8 +26,8 @@ func (s *subCategHandler) CreateSubCateg(w http.ResponseWriter, r *http.Request)
 		Name        string `json:"name"`
 		MainCategID int64  `json:"main_category_id"`
 	}
-	if err := jsutil.ReadJson(w, r, &input); err != nil {
-		logger.Error("jsutil.ReadJSON failed", "package", "handler", "err", err)
+	if err := jsonutil.ReadJson(w, r, &input); err != nil {
+		logger.Error("jsonutil.ReadJSON failed", "package", "handler", "err", err)
 		errutil.BadRequestResponse(w, r, err)
 		return
 	}
@@ -68,17 +68,17 @@ func (s *subCategHandler) GetAllSubCateg(w http.ResponseWriter, r *http.Request)
 	respData := map[string]interface{}{
 		"sub_categories": categs,
 	}
-	if err := jsutil.WriteJSON(w, http.StatusOK, respData, nil); err != nil {
-		logger.Error("jsutil.WriteJSON failed", "package", "handler", "err", err)
+	if err := jsonutil.WriteJSON(w, http.StatusOK, respData, nil); err != nil {
+		logger.Error("jsonutil.WriteJSON failed", "package", "handler", "err", err)
 		errutil.ServerErrorResponse(w, r, err)
 		return
 	}
 }
 
 func (s *subCategHandler) GetByMainCategID(w http.ResponseWriter, r *http.Request) {
-	id, err := jsutil.ReadID(r)
+	id, err := jsonutil.ReadID(r)
 	if err != nil {
-		logger.Error("jsutil.ReadID failed", "package", "handler", "err", err)
+		logger.Error("jsonutil.ReadID failed", "package", "handler", "err", err)
 		errutil.BadRequestResponse(w, r, err)
 		return
 	}
@@ -94,17 +94,17 @@ func (s *subCategHandler) GetByMainCategID(w http.ResponseWriter, r *http.Reques
 	respData := map[string]interface{}{
 		"sub_categories": categs,
 	}
-	if err := jsutil.WriteJSON(w, http.StatusOK, respData, nil); err != nil {
-		logger.Error("jsutil.WriteJSON failed", "package", "handler", "err", err)
+	if err := jsonutil.WriteJSON(w, http.StatusOK, respData, nil); err != nil {
+		logger.Error("jsonutil.WriteJSON failed", "package", "handler", "err", err)
 		errutil.ServerErrorResponse(w, r, err)
 		return
 	}
 }
 
 func (s *subCategHandler) UpdateSubCateg(w http.ResponseWriter, r *http.Request) {
-	id, err := jsutil.ReadID(r)
+	id, err := jsonutil.ReadID(r)
 	if err != nil {
-		logger.Error("jsutil.ReadID failed", "package", "handler", "err", err)
+		logger.Error("jsonutil.ReadID failed", "package", "handler", "err", err)
 		errutil.BadRequestResponse(w, r, err)
 		return
 	}
@@ -112,8 +112,8 @@ func (s *subCategHandler) UpdateSubCateg(w http.ResponseWriter, r *http.Request)
 	var input struct {
 		Name string `json:"name"`
 	}
-	if err := jsutil.ReadJson(w, r, &input); err != nil {
-		logger.Error("jsutil.ReadJSON failed", "package", "handler", "err", err)
+	if err := jsonutil.ReadJson(w, r, &input); err != nil {
+		logger.Error("jsonutil.ReadJSON failed", "package", "handler", "err", err)
 		errutil.BadRequestResponse(w, r, err)
 		return
 	}
@@ -143,9 +143,9 @@ func (s *subCategHandler) UpdateSubCateg(w http.ResponseWriter, r *http.Request)
 }
 
 func (s *subCategHandler) DeleteSubCateg(w http.ResponseWriter, r *http.Request) {
-	id, err := jsutil.ReadID(r)
+	id, err := jsonutil.ReadID(r)
 	if err != nil {
-		logger.Error("jsutil.ReadID failed", "package", "handler", "err", err)
+		logger.Error("jsonutil.ReadID failed", "package", "handler", "err", err)
 		errutil.BadRequestResponse(w, r, err)
 		return
 	}
@@ -156,8 +156,8 @@ func (s *subCategHandler) DeleteSubCateg(w http.ResponseWriter, r *http.Request)
 		return
 	}
 
-	if err := jsutil.WriteJSON(w, http.StatusOK, nil, nil); err != nil {
-		logger.Error("jsutil.WriteJSON failed", "package", "handler", "err", err)
+	if err := jsonutil.WriteJSON(w, http.StatusOK, nil, nil); err != nil {
+		logger.Error("jsonutil.WriteJSON failed", "package", "handler", "err", err)
 		errutil.ServerErrorResponse(w, r, err)
 		return
 	}
